gallery and site update

+++ src/pages/document/documents/DocumentHistoryViewForm.js
@@ -481,6 +481,7 @@ const handleNewFile = async () => {
                       updatedByFullName: version?.updatedBy?.name || '',
                       updatedIP: version?.updatedIP || '',
                     }
+
                  return (
                   <Grid container key={index}>
                     <Grid container sx={{ pt: '2rem' }} mb={1}>
@@ -501,29 +502,65 @@ const handleNewFile = async () => {
                         xl: 'repeat(8, 1fr)',
                       }}
                     >
-                      {version?.files?.map((file, _index) => 
-                        <DocumentGalleryItem key={file?._id} image={{
-                          thumbnail: `data:image/png;base64, ${file.thumbnail}`,
-                          src: `data:image/png;base64, ${file.thumbnail}`,
-                          downloadFilename: `${file?.name}.${file?.extension}`,
-                          name: file?.name,
-                          category: file?.docCategory?.name,
-                          fileType: file?.fileType,
-                          extension: file?.extension,
-                          isLoaded: false,
-                          id: file?._id,
-                          width: '100%',
-                          height: '100%',
-                        }} isLoading={isLoading} 
-                        onOpenLightbox={()=> handleOpenLightbox(_index)}
-                        onDownloadFile={()=> handleDownloadFile(documentHistory._id, version._id, file._id, file?.name, file?.extension)}
-                        onDeleteFile={()=> handleDeleteFile(documentHistory._id, version._id, file._id)}
-                        onOpenFile={()=> handleOpenFile(documentHistory._id, version._id, file._id, file)}
-                        toolbar
-                        
-                        />
-                        
-                      )}
+                      { version?.files?.map((file, _index) =>{
+                        if (file.fileType.startsWith('image')) {
+                          return (
+                            <DocumentGalleryItem
+                              key={file?._id}
+                              image={{
+                                thumbnail: `data:image/png;base64, ${file.thumbnail}`,
+                                src: `data:image/png;base64, ${file.thumbnail}`,
+                                downloadFilename: `${file?.name}.${file?.extension}`,
+                                name: file?.name,
+                                category: file?.docCategory?.name,
+                                fileType: file?.fileType,
+                                extension: file?.extension,
+                                isLoaded: false,
+                                id: file?._id,
+                                width: '100%',
+                                height: '100%',
+                              }}
+                              isLoading={isLoading} 
+                              onOpenLightbox={() => handleOpenLightbox(index)}
+                              onDownloadFile={() => handleDownloadFile(documentHistory._id, version._id, file._id, file?.name, file?.extension)}
+                              onDeleteFile={() => handleDeleteFile(documentHistory._id, version._id, file._id)}
+                              onOpenFile={() => handleOpenFile(documentHistory._id, version._id, file._id, file)}
+                              toolbar
+                            />
+                          );
+                        }
+                        return null;
+                      })}
+
+                      {version?.files?.map((file, _index) =>{
+                        if (!file.fileType.startsWith('image')) {
+                          return (
+                            <DocumentGalleryItem
+                              key={file?._id}
+                              image={{
+                                thumbnail: `data:image/png;base64, ${file.thumbnail}`,
+                                src: `data:image/png;base64, ${file.thumbnail}`,
+                                downloadFilename: `${file?.name}.${file?.extension}`,
+                                name: file?.name,
+                                category: file?.docCategory?.name,
+                                fileType: file?.fileType,
+                                extension: file?.extension,
+                                isLoaded: false,
+                                id: file?._id,
+                                width: '100%',
+                                height: '100%',
+                              }}
+                              isLoading={isLoading} 
+                              // onOpenLightbox={() => handleOpenLightbox(index)}
+                              onDownloadFile={() => handleDownloadFile(documentHistory._id, version._id, file._id, file?.name, file?.extension)}
+                              onDeleteFile={() => handleDeleteFile(documentHistory._id, version._id, file._id)}
+                              onOpenFile={() => handleOpenFile(documentHistory._id, version._id, file._id, file)}
+                              toolbar
+                            />
+                          );
+                        }
+                        return null;
+                      })}
 
                       {index === 0 && !defaultValues.isArchived && (<ThumbnailDocButton onClick={handleNewFile}/>)}
                     </Box>
